Expose metadata limits and annotation info to filter stack
