+++ BUILD
@@ -4245,6 +4245,7 @@ grpc_cc_library(
         "//src/core:hpack_constants",
         "//src/core:match",
         "//src/core:metadata_batch",
+        "//src/core:metadata_info",
         "//src/core:parsed_metadata",
         "//src/core:random_early_detection",
         "//src/core:slice",
@@ -4422,6 +4423,7 @@ grpc_cc_library(
         "//src/core:max_concurrent_streams_policy",
         "//src/core:memory_quota",
         "//src/core:metadata_batch",
+        "//src/core:metadata_info",
         "//src/core:ping_abuse_policy",
         "//src/core:ping_callbacks",
         "//src/core:ping_rate_policy",
