+++ build_autogenerated.yaml
@@ -1145,6 +1145,7 @@ libs:
   - src/core/lib/transport/metadata.h
   - src/core/lib/transport/metadata_batch.h
   - src/core/lib/transport/metadata_compression_traits.h
+  - src/core/lib/transport/metadata_info.h
   - src/core/lib/transport/parsed_metadata.h
   - src/core/lib/transport/simple_slice_based_metadata.h
   - src/core/lib/transport/status_conversion.h
@@ -1943,6 +1944,7 @@ libs:
   - src/core/lib/transport/message.cc
   - src/core/lib/transport/metadata.cc
   - src/core/lib/transport/metadata_batch.cc
+  - src/core/lib/transport/metadata_info.cc
   - src/core/lib/transport/parsed_metadata.cc
   - src/core/lib/transport/status_conversion.cc
   - src/core/lib/transport/tcp_connect_handshaker.cc
@@ -2605,6 +2607,7 @@ libs:
   - src/core/lib/transport/metadata.h
   - src/core/lib/transport/metadata_batch.h
   - src/core/lib/transport/metadata_compression_traits.h
+  - src/core/lib/transport/metadata_info.h
   - src/core/lib/transport/parsed_metadata.h
   - src/core/lib/transport/simple_slice_based_metadata.h
   - src/core/lib/transport/status_conversion.h
@@ -3023,6 +3026,7 @@ libs:
   - src/core/lib/transport/message.cc
   - src/core/lib/transport/metadata.cc
   - src/core/lib/transport/metadata_batch.cc
+  - src/core/lib/transport/metadata_info.cc
   - src/core/lib/transport/parsed_metadata.cc
   - src/core/lib/transport/status_conversion.cc
   - src/core/lib/transport/tcp_connect_handshaker.cc
