+++ config.w32
@@ -762,6 +762,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\transport\\message.cc " +
     "src\\core\\lib\\transport\\metadata.cc " +
     "src\\core\\lib\\transport\\metadata_batch.cc " +
+    "src\\core\\lib\\transport\\metadata_info.cc " +
     "src\\core\\lib\\transport\\parsed_metadata.cc " +
     "src\\core\\lib\\transport\\status_conversion.cc " +
     "src\\core\\lib\\transport\\tcp_connect_handshaker.cc " +
