+++ grpc.gyp
@@ -983,6 +983,7 @@
         'src/core/lib/transport/message.cc',
         'src/core/lib/transport/metadata.cc',
         'src/core/lib/transport/metadata_batch.cc',
+        'src/core/lib/transport/metadata_info.cc',
         'src/core/lib/transport/parsed_metadata.cc',
         'src/core/lib/transport/status_conversion.cc',
         'src/core/lib/transport/tcp_connect_handshaker.cc',
@@ -1485,6 +1486,7 @@
         'src/core/lib/transport/message.cc',
         'src/core/lib/transport/metadata.cc',
         'src/core/lib/transport/metadata_batch.cc',
+        'src/core/lib/transport/metadata_info.cc',
         'src/core/lib/transport/parsed_metadata.cc',
         'src/core/lib/transport/status_conversion.cc',
         'src/core/lib/transport/tcp_connect_handshaker.cc',
