+++ package.xml
@@ -1813,6 +1813,8 @@
     <file baseinstalldir="/" name="src/core/lib/transport/metadata_batch.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/transport/metadata_batch.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/transport/metadata_compression_traits.h" role="src" />
+    <file baseinstalldir="/" name="src/core/lib/transport/metadata_info.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/lib/transport/metadata_info.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/transport/parsed_metadata.cc" role="src" />
     <file baseinstalldir="/" name="src/core/lib/transport/parsed_metadata.h" role="src" />
     <file baseinstalldir="/" name="src/core/lib/transport/simple_slice_based_metadata.h" role="src" />
